requirements.txt
asgiref==3.4.1
cloudinary==1.28.1
dj-database-url==0.5.0
dj3-cloudinary-storage==0.0.6
Django==3.2
django-allauth==0.47.0
gunicorn==20.1.0
oauthlib==3.1.1
psycopg2==2.9.3
PyJWT==2.3.0
python3-openid==3.2.0
pytz==2021.3
requests-oauthlib==1.3.0
slugify==0.0.1
sqlparse==0.4.2
